Collect integration tests
